NSB over-voltage investigation
Following good discussion with Jason on the next investigation, I've modeled the effect of decreasing over-voltage with increasing NSB, which Jon L. has shown decreases SiPM response to a laser pulse by 0.016% per MHz increase in NSB.

The method I've coded up is :

   1) initial pedestal subtraction at 50 MHz NSB
   2) Increase NSB in a linear fashion 50 - 1000 MHz
   3) Intersperce this NSB increase with flasher calibration runs where I measure the combined pedestal corrected flasher illumination + NSB ( "measured illumination" in attached plots), with each run having some number of flashes and flash rate
    4) I simulate the decreased silicon response by dropping the flasher illumination, (so called "true illumination") depending on the NSB at the time of the flash
    5) I determine the root mean square error (RMSE) to quantify the mis-calibration between "true" and "measured"

My code also now uses  the common NSB pedestal subtraction and charge extraction routines in the simulator which should supresses NSB in relation to flash signal as it considers a narrow 20 ns window rather than whole waveform integration.

I attach some plots which show how "measured" and "true" vary with different flash rates, NSB rate of change, and number of flashes in each calibration run (which are ten minutes apart)

The plots show a largely linear relationship between measured and true (as expected) whose linearity improves with number of flashes. The  NSB rate of change plot is less useful as with increasing rate of change, we have fewer calibration runs and hence points on the plot, before getting to 1 GHz NSB.

I also generated some output log files (attached) with my stats for these runs and the variable parameters for each run.  In the files "MPRC" is "measured" and "TPRC" is "true". The take home message is that the mis-calibration RMSE varies from approximately 4 to 14 p.e. which for an initial flash of 100 p.e. can be taken as 4 to 14 % miscalibration which is within the stated requirement of 20 % . This is pretty much as expected from Jon L. figures (1000 MHz * 0.016 = 16 %). Increasing the number of flashes also takes the standard deviation on the measured calibration runs down from 7 p.e to <2 p.e. for 10 to 100 flashes in each calibration run respectively.  Going to 200 flashes (files not attached) only takes standard dev down to around 1 p.e. so diminishing return. 

I've also coded up a sin wave NSB generator so we can sample at any time within a max min NSB range with a given rise time and that is the next thing I will try to make sure the stats are not affected much by the level of the start NSB pedestal. (This will improve on the linear NSB generator which stops at around 1 GHz and only allows me so much time to place the calibration runs in as I allude too above).